connector/requirements.txt
pulsar-client[avro]>=3.0.0
requests>=2.28.0
